pyproject.toml
[tool.poetry]
name = "raund"
version = "1.0.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
black = "^23.9.1"
isort = "^5.9.3"
mypy = "^1.5.1"
jupyterlab = "^4.0.6"
SQLAlchemy = "^1.4.29"
s3fs = "^2023.9.2"
pandas = "^2.1.1"
boto3 = "1.28.17"

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.5.1"
slack-bolt = "^1.14.0"
slack-sdk = "^3.16.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
target-version = ['py38', 'py39']
line-length = 120
exclude = """
/(
    .git
    | .github
    | .venv
    | .mypy_cache
    | .pytest_cache
    | cdkv2
    | data
    | docs
    | htmlcov
    | notebook
    | frontend
    | cdk.out
    | node_modules
/)
"""

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true